Moved test reader to a separate directory
This makes it easier to validate the tool's behavior separately from the diff processor
