+++ .github/workflows/unit-test-tools.yml
@@ -104,3 +104,23 @@ jobs:
         run: |
           cd tools/template-check
           go test -v ./...
+
+  test-reader:
+    runs-on: ubuntu-22.04
+    steps:
+      - uses: actions/checkout@b4ffde65f46336ab88eb53be808477a3936bae11 # v4.1.2
+
+      - name: Set up Go
+        uses: actions/setup-go@0c52d547c9bc32b1aa3301fd7a9cb496313a4491 # v5.0.0
+        with:
+          go-version: '^1.21.0'
+
+      - name: Build test-reader
+        run: |
+          cd tools/test-reader
+          go build
+
+      - name: Test test-reader
+        run: |
+          cd tools/test-reader
+          go test -v ./...
