--- tools/diff-processor/reader/testdata/service/cross_file_2_test.go
@@ -1,36 +1,0 @@
-package service_test
-
-import (
-	"google/provider/new/google-beta/acctest"
-	"testing"
-
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-)
-
-func testAccCrossFile2(t *testing.T) {
-	VcrTest(t, resource.TestCase{
-		Steps: []resource.TestStep{
-			{
-				Config: testAccCrossFileConfig2(),
-			},
-		},
-	})
-}
-
-func testAccCrossFileConfig1() string {
-	return acctest.Nprintf(`
-resource "serial_resource" "resource" {
-  field_one = "value-one"
-}
-`, context)
-}
-
-func testAccCrossFileConfig2() string {
-	return acctest.Nprintf(`
-resource "serial_resource" "resource" {
-  field_two {
-    field_three = "value-two"
-  }
-}
-`, context)
-}
