--- tools/diff-processor/reader/testdata/service/uncovered_resource_test.go
@@ -1,28 +1,0 @@
-package service_test
-
-import (
-	"testing"
-
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-	"github.com/hashicorp/terraform-provider-google-beta/google-beta/acctest"
-)
-
-func TestAccUncoveredResource(t *testing.T) {
-	VcrTest(t, resource.TestCase{
-		Steps: []resource.TestStep{
-			{
-				Config: testAccUncoveredResource(),
-			},
-		},
-	})
-}
-
-func testAccUncoveredResource() string {
-	return acctest.Nprintf(`
-resource "uncovered_resource" "resource" {
-  field_two {
-    field_three = "value-two"
-  }
-}
-`, context)
-}
