+++ tools/test-reader/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"github.com/GoogleCloudPlatform/magic-modules/tools/test-reader/cmd"
+)
+
+func main() {
+	cmd.Execute()
+}
