+++ tools/test-reader/reader/testdata/service/config_variable_test.go
@@ -1,0 +1,23 @@
+package service_test
+
+import (
+	"testing"
+
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
+)
+
+var testConfigVariable = `
+resource "config_variable" "basic" {
+  field_one = "value-one"
+}
+`
+
+func TestAccSqlDatabaseInstance_basicInferredName(t *testing.T) {
+	VcrTest(t, resource.TestCase{
+		Steps: []resource.TestStep{
+			{
+				Config: testConfigVariable,
+			},
+		},
+	})
+}
